requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.23
cryptography==41.0.4
paramiko==3.3.1
scikit-learn==1.3.0
pandas==2.1.1
matplotlib==3.8.0
python-dotenv==1.0.0
pyOpenSSL==23.2.0
flask-sqlalchemy==3.1.1
flask-login==0.6.2
flask-wtf==1.2.1
flask-migrate==4.0.5
Werkzeug==2.3.7 